requirements.txt
pandas
numpy
jupyterlab
python-dateutil
fastapi
uvicorn[standard]